src/knowledge-graph-backend/requirements.txt
aiohappyeyeballs==2.4.2
aiohttp==3.10.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
azure-core==1.31.0
azure-identity==1.18.0
azure-storage-blob==12.23.1
black==24.8.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
distro==1.9.0
faiss-cpu==1.8.0.post1
fastapi==0.115.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.10
isodate==0.6.1
jiter==0.5.0
msal==1.31.0
msal-extensions==1.2.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
openai==1.50.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
portalocker==2.10.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
python-multipart==0.0.10
# pywin32==306
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
watchfiles==0.24.0
websockets==13.1
yarl==1.13.0
